# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PinView Golf Social Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  username: string;
  displayName: string;
  bio?: string;
  avatarUrl?: string;
  handicap?: number | null;
  homeCourse?: string | null;
  followersCount: number;
  followingCount: number;
  isBlocked?: boolean;
  postsCount: number;
  isFollowing: boolean;
  createdAt: string;
}

export type UserProfile = User;

export interface UserCard {
  id: string;
  username: string;
  displayName: string;
  avatarUrl?: string;
  handicap?: number | null;
  homeCourse?: string | null;
  followersCount: number;
  isFollowing: boolean;
}

export interface UserStats {
  totalShots: number;
  totalLikes: number;
  avgDistance?: number | null;
  longestShot?: number | null;
  topShotType?: string | null;
  topClub?: string | null;
}

export type PostShotShape =
  | (typeof PostShotShape)[keyof typeof PostShotShape]
  | null;

export const PostShotShape = {
  Straight: "Straight",
  Draw: "Draw",
  Fade: "Fade",
  Hook: "Hook",
  Slice: "Slice",
} as const;

export interface Post {
  id: string;
  author: UserCard;
  caption?: string | null;
  videoUrl?: string | null;
  thumbnailUrl?: string | null;
  course?: string | null;
  holeNumber?: number | null;
  club?: string | null;
  distance?: number | null;
  shotShape?: PostShotShape;
  shotType?: string | null;
  tags: string[];
  likesCount: number;
  commentsCount: number;
  savesCount: number;
  isLiked: boolean;
  isSaved: boolean;
  createdAt: string;
}

export interface Comment {
  id: string;
  author: UserCard;
  text: string;
  createdAt: string;
}

export interface CourseCard {
  name: string;
  location?: string | null;
  postsCount: number;
  thumbnailUrl?: string | null;
}

export interface PostsPage {
  posts: Post[];
  nextCursor?: string | null;
  hasMore: boolean;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  like: "like",
  comment: "comment",
  follow: "follow",
  featured: "featured",
  trending: "trending",
} as const;

export interface Notification {
  id: string;
  type: NotificationType;
  actor: UserCard;
  post?: Post | null;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface NotificationsPage {
  notifications: Notification[];
  nextCursor?: string | null;
  hasMore: boolean;
}

export interface UnreadCount {
  count: number;
}

export interface SearchResults {
  users: UserCard[];
  posts: Post[];
  courses: CourseCard[];
}

export interface RegisterBody {
  username: string;
  displayName: string;
  email: string;
  password: string;
  handicap?: number | null;
  homeCourse?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
}

export interface UpdateUserBody {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  handicap?: number | null;
  homeCourse?: string | null;
}

export type CreatePostBodyShotShape =
  | (typeof CreatePostBodyShotShape)[keyof typeof CreatePostBodyShotShape]
  | null;

export const CreatePostBodyShotShape = {
  Straight: "Straight",
  Draw: "Draw",
  Fade: "Fade",
  Hook: "Hook",
  Slice: "Slice",
} as const;

export interface CreatePostBody {
  caption?: string | null;
  videoUrl?: string | null;
  thumbnailUrl?: string | null;
  course?: string | null;
  holeNumber?: number | null;
  club?: string | null;
  distance?: number | null;
  shotShape?: CreatePostBodyShotShape;
  shotType?: string | null;
  tags?: string[];
}

export interface CreateCommentBody {
  text: string;
}

export type GetUserPostsParams = {
  cursor?: string;
  limit?: number;
};

export type GetUserSavedParams = {
  limit?: number;
};

export type GetFeedParams = {
  cursor?: string;
  limit?: number;
};

export type GetTrendingPostsParams = {
  limit?: number;
};

export type SearchParams = {
  q: string;
  type?: SearchType;
};

export type SearchType = (typeof SearchType)[keyof typeof SearchType];

export const SearchType = {
  all: "all",
  users: "users",
  posts: "posts",
  courses: "courses",
} as const;

export type GetNotificationsParams = {
  cursor?: string;
  limit?: number;
};