# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PinView Golf Social Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new golf shot post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Liked

    delete:
      operationId: unlikePost
      tags: [posts]
      summary: Unlike a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unliked
  /posts/{postId}/save:
    post:
      operationId: savePost
      tags: [posts]
      summary: Save a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Saved

    delete:
      operationId: unsavePost
      tags: [posts]
      summary: Unsave a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unsaved
  /posts/{postId}/comments:
    get:
      operationId: getPostComments
      tags: [posts]
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [posts]
      summary: Comment on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /posts/{postId}/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [posts]
      summary: Delete own comment
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # Feed
...
components:
  schemas:
...
    UserCard:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        handicap:
          type: number
          nullable: true
        homeCourse:
          type: string
          nullable: true
        followersCount:
          type: integer
        isFollowing:
          type: boolean
      required: [id, username, displayName, followersCount, isFollowing]
...
    Post:
      type: object
      properties:
        id:
          type: string
        author:
          $ref: "#/components/schemas/UserCard"
        caption:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        course:
          type: string
          nullable: true
        holeNumber:
          type: integer
          nullable: true
        club:
          type: string
          nullable: true
        distance:
          type: integer
          nullable: true
        shotShape:
          type: string
          enum: [Straight, Draw, Fade, Hook, Slice]
          nullable: true
        shotType:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        savesCount:
          type: integer
        isLiked:
          type: boolean
        isSaved:
          type: boolean
        createdAt:
          type: string
      required: [id, author, tags, likesCount, commentsCount, savesCount, isLiked, isSaved, createdAt]
    Comment:
      type: object
      properties:
        id:
          type: string
        author:
          $ref: "#/components/schemas/UserCard"
        text:
          type: string
        createdAt:
          type: string
      required: [id, author, text, createdAt]
...
    CreatePostBody:
      type: object
      properties:
        caption:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        course:
          type: string
          nullable: true
        holeNumber:
          type: integer
          nullable: true
        club:
          type: string
          nullable: true
        distance:
          type: integer
          nullable: true
        shotShape:
          type: string
          enum: [Straight, Draw, Fade, Hook, Slice]
          nullable: true
        shotType:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
    CreateCommentBody:
      type: object
      properties:
        text:
          type: string
      required: [text]